# server/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: felixtj.dietrich@tum.de
    name: Felix T.J. Dietrich
  description: API documentation for the Hephaestus application server.
  license:
    name: MIT License
    url: https://github.com/ls1intum/Hephaestus/blob/develop/LICENSE
  title: Hephaestus API
  version: 0.0.0-development
servers:
- description: Default Server URL
  url: /
security:
- bearerAuth: []
...
paths:
...
  /workspaces/{workspaceSlug}:
    delete:
      operationId: purgeWorkspace
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      responses:
        "204":
          description: Workspace purged
      summary: Purge (soft delete) a workspace
      tags:
      - workspace
    get:
      operationId: getWorkspace
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace returned
      security: []
      summary: Fetch a workspace by slug
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/features:
    patch:
      operationId: updateFeatures
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceFeaturesRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace updated
      summary: Update workspace feature flags
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/leaderboard-digest:
    patch:
      operationId: updateLeaderboardDigest
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaderboardDigestRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace updated
      summary: Update the whole weekly leaderboard digest config (schedule + notifications)
        atomically
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/league/reset:
    put:
      operationId: resetAndRecalculateLeagues
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      responses:
        "200":
          description: OK
      summary: Reset and recalculate workspace leagues
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/notifications:
    patch:
      operationId: updateNotifications
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceNotificationsRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace updated
      summary: Update leaderboard notification preferences
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/public-visibility:
    patch:
      operationId: updatePublicVisibility
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspacePublicVisibilityRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace updated
      summary: Toggle public visibility for a workspace
      tags:
      - workspace
  /workspaces/{workspaceSlug}/repositories:
    delete:
      operationId: removeRepositoryToMonitor
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: query
        name: nameWithOwner
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      summary: Remove a repository from a workspace monitor list
      tags:
      - workspace
    get:
      operationId: getRepositoriesToMonitor
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
          description: Repository list
      security: []
      summary: List repositories monitored by a workspace
      tags:
      - workspace
    post:
      operationId: addRepositoryToMonitor
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: query
        name: nameWithOwner
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      summary: Add a repository to a workspace monitor list
      tags:
      - workspace
  /workspaces/{workspaceSlug}/schedule:
    patch:
      operationId: updateSchedule
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceScheduleRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace updated
      summary: Update leaderboard schedule configuration
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/slug:
    patch:
      operationId: renameSlug
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameWorkspaceSlugRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace renamed
      summary: Rename workspace slug and create redirect
      tags:
      - workspace
  /workspaces/{workspaceSlug}/status:
    patch:
      operationId: updateStatus
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceStatusRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace updated
      summary: Update workspace lifecycle status
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/teams/{teamId}/labels/{labelId}:
    delete:
      operationId: removeLabelFromTeam
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: labelId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInfo"
          description: OK
      summary: Remove a repository label from a team
      tags:
      - workspace
  /workspaces/{workspaceSlug}/teams/{teamId}/labels/{repositoryId}/{label}:
    post:
      operationId: addLabelToTeam
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      - in: path
        name: teamId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: repositoryId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: label
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInfo"
          description: OK
      summary: Add a repository label to a team
      tags:
      - workspace
...
  /workspaces/{workspaceSlug}/token:
    patch:
      operationId: updateToken
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceTokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
          description: Workspace updated
      summary: Update workspace Personal Access Token
      tags:
      - workspace
  /workspaces/{workspaceSlug}/users:
    get:
      operationId: getUsersWithTeams
      parameters:
      - description: Workspace slug
        in: path
        name: workspaceSlug
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserTeams"
          description: OK
      security: []
      summary: List workspace users and the teams they belong to
      tags:
      - workspace
...
components:
  schemas:
...
    RenameWorkspaceSlugRequest:
      type: object
      description: Request to rename a workspace's URL slug
      properties:
        newSlug:
          type: string
          description: New URL-friendly identifier for the workspace
          example: new-workspace-slug
          minLength: 1
          pattern: "^[a-z0-9][a-z0-9-]{2,50}$"
      required:
      - newSlug
...
    UpdateLeaderboardDigestRequest:
      type: object
      description: Request to update the entire weekly leaderboard digest configuration
        atomically
      properties:
        channelId:
          type: string
          description: Slack channel ID for notifications
          example: C01234567AB
          pattern: "^[CG][A-Z0-9]{8,}$"
        day:
          type: integer
          format: int32
          description: "Day of week (1=Monday, 7=Sunday)"
          example: 1
          maximum: 7
          minimum: 1
        enabled:
          type: boolean
          description: Whether leaderboard notifications are enabled
        team:
          type: string
          description: Team name for filtering leaderboard notifications
        time:
          type: string
          description: Time in 24-hour format (HH:mm)
          example: 09:00
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
      required:
      - day
      - time
...
    UpdateWorkspaceFeaturesRequest:
      type: object
      description: Request to update workspace feature flags. Null fields are left
        unchanged.
      properties:
        achievementsEnabled:
          type: boolean
          description: Enable the achievements system
        leaderboardEnabled:
          type: boolean
          description: Enable the leaderboard ranking page
        leaguesEnabled:
          type: boolean
          description: Enable league tiers and rankings
        mentorEnabled:
          type: boolean
          description: Enable the Pi mentor chat feature
        practiceReviewAutoTriggerEnabled:
          type: boolean
          description: Enable automatic practice reviews triggered by PR events
        practiceReviewManualTriggerEnabled:
          type: boolean
          description: Enable manual practice reviews triggered via bot command
        practicesEnabled:
          type: boolean
          description: Enable the practice review feature
        progressionEnabled:
          type: boolean
          description: Enable the league/progression system
    UpdateWorkspaceNotificationsRequest:
      type: object
      description: Request to update leaderboard notification settings
      properties:
        channelId:
          type: string
          description: Slack channel ID for notifications
          example: C01234567AB
          pattern: "^[CG][A-Z0-9]{8,}$"
        enabled:
          type: boolean
          description: Whether leaderboard notifications are enabled
        team:
          type: string
          description: Team name for filtering leaderboard notifications
    UpdateWorkspacePublicVisibilityRequest:
      type: object
      description: Request to update workspace public visibility setting
      properties:
        isPubliclyViewable:
          type: boolean
          description: Whether the workspace should be publicly viewable without authentication
      required:
      - isPubliclyViewable
    UpdateWorkspaceScheduleRequest:
      type: object
      description: Request to update the leaderboard notification schedule
      properties:
        day:
          type: integer
          format: int32
          description: "Day of week (1=Monday, 7=Sunday)"
          example: 1
          maximum: 7
          minimum: 1
        time:
          type: string
          description: Time in 24-hour format (HH:mm)
          example: 09:00
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
      required:
      - day
      - time
    UpdateWorkspaceStatusRequest:
      type: object
      description: Request to update the workspace lifecycle status
      properties:
        status:
          type: string
          description: "New lifecycle status (PENDING, ACTIVE, ARCHIVED)"
          enum:
          - ACTIVE
          - SUSPENDED
          - PURGED
      required:
      - status
    UpdateWorkspaceTokenRequest:
      type: object
      description: Request to update the workspace's GitHub Personal Access Token
      properties:
        personalAccessToken:
          type: string
          description: GitHub Personal Access Token for API access
          minLength: 1
      required:
      - personalAccessToken
...
    Workspace:
      type: object
      description: Complete workspace information including configuration and settings
      properties:
        accountLogin:
          type: string
          description: Git provider account login associated with this workspace
        achievementsEnabled:
          type: boolean
          description: Whether the achievements system is enabled
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the workspace was created
        displayName:
          type: string
          description: Human-readable name of the workspace
        gitlabWebhookRegistered:
          type: boolean
          description: Whether a GitLab webhook has been auto-registered for this
            workspace
        hasPersonalAccessToken:
          type: boolean
          description: Whether a Personal Access Token is configured
        hasSlackToken:
          type: boolean
          description: Whether Slack token is configured
        id:
          type: integer
          format: int64
          description: Unique identifier of the workspace
        installationId:
          type: integer
          format: int64
          description: "GitHub App installation ID, if linked"
        installationLinkedAt:
          type: string
          format: date-time
          description: Timestamp when the GitHub App installation was linked
        isPubliclyViewable:
          type: boolean
          description: Whether the workspace is publicly viewable without authentication
        kind:
          type: string
          description: Integration kind backing this workspace (GITHUB or GITLAB)
        leaderboardEnabled:
          type: boolean
          description: Whether the leaderboard is enabled
        leaderboardNotificationChannelId:
          type: string
          description: Slack channel ID for leaderboard notifications
        leaderboardNotificationEnabled:
          type: boolean
          description: Whether leaderboard notifications are enabled
        leaderboardNotificationTeam:
          type: string
          description: Team name for leaderboard notifications
        leaderboardScheduleDay:
          type: integer
          format: int32
          description: "Day of week for leaderboard notifications (1=Monday, 7=Sunday)"
          example: 1
        leaderboardScheduleTime:
          type: string
          description: Time for leaderboard notifications in HH:mm format
          example: 09:00
        leaguesEnabled:
          type: boolean
          description: Whether league tiers and rankings are enabled
        mentorEnabled:
          type: boolean
          description: Whether the Pi mentor chat feature is enabled
        practiceReviewAutoTriggerEnabled:
          type: boolean
          description: Whether automatic practice reviews triggered by PR events are
            enabled
        practiceReviewManualTriggerEnabled:
          type: boolean
          description: Whether manual practice reviews triggered via bot command are
            enabled
        practicesEnabled:
          type: boolean
          description: Whether the practice review feature is enabled
        progressionEnabled:
          type: boolean
          description: Whether the league/progression system is enabled
        providerType:
          type: string
          description: High-level git provider type for the workspace's SCM connection
            (null if none bound)
          enum:
          - GITHUB
          - GITLAB
          - SLACK
          - OUTLINE
        serverUrl:
          type: string
          description: Custom server URL for self-hosted instances (null for cloud
            defaults)
        slackConnectionId:
          type: integer
          format: int64
          description: "ID of the active Slack connection, if any — addresses PATCH\
            \ /connections/{id}/status"
        status:
          type: string
          description: "Current lifecycle status of the workspace (PENDING, ACTIVE,\
            \ ARCHIVED)"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the workspace was last updated
        workspaceSlug:
          type: string
          description: URL-friendly identifier for the workspace
          example: my-workspace
      required:
      - accountLogin
      - achievementsEnabled
      - createdAt
      - displayName
      - gitlabWebhookRegistered
      - hasPersonalAccessToken
      - hasSlackToken
      - id
      - isPubliclyViewable
      - leaderboardEnabled
      - leaguesEnabled
      - mentorEnabled
      - practiceReviewAutoTriggerEnabled
      - practiceReviewManualTriggerEnabled
      - practicesEnabled
      - progressionEnabled
      - status
      - updatedAt
      - workspaceSlug
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Hephaestus-native JWT bearer authentication. The SPA normally authenticates
        via the `__Host-HEPHAESTUS_AT` session cookie; this scheme documents the equivalent
        bearer token.
      scheme: bearer
      type: http